requirements.txt
# ============================================================
# IB Trading System - Python Dependencies
# ============================================================

# ------------------------------------------------------------
# IB 連接
# ------------------------------------------------------------
ib_insync>=0.9.86

# ------------------------------------------------------------
# 數據處理
# ------------------------------------------------------------
pandas>=2.0.0
numpy>=1.24.0

# ------------------------------------------------------------
# 數據庫
# ------------------------------------------------------------
sqlalchemy>=2.0.0
duckdb>=0.9.0

# ------------------------------------------------------------
# 異步支援
# ------------------------------------------------------------
aiofiles>=23.0.0

# ------------------------------------------------------------
# 排程
# ------------------------------------------------------------
apscheduler>=3.10.0

# ------------------------------------------------------------
# 日誌
# ------------------------------------------------------------
loguru>=0.7.0

# ------------------------------------------------------------
# 環境變數
# ------------------------------------------------------------
python-dotenv>=1.0.0

# ------------------------------------------------------------
# 通知服務
# ------------------------------------------------------------
python-telegram-bot>=20.0
aiosmtplib>=3.0.0

# ------------------------------------------------------------
# 技術指標
# ------------------------------------------------------------
ta>=0.11.0
pandas-ta>=0.3.14b

# ------------------------------------------------------------
# 時區處理
# ------------------------------------------------------------
pytz>=2024.1

# ------------------------------------------------------------
# 性能監控
# ------------------------------------------------------------
psutil>=5.9.0

# ------------------------------------------------------------
# 測試工具
# ------------------------------------------------------------
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# ------------------------------------------------------------
# 開發工具 (可選)
# ------------------------------------------------------------
# black>=23.0.0
# ruff>=0.1.0
# mypy>=1.0.0